Index A/V playlist data from Kaltura
In preparation for migrating to UniversalViewer, we need to index
some data for audio and video playlists that is stored in Kaltura. Each
playlist item must have a duration and a name. We'll store that as we
ingest documents, but this change also provides a means to backfill this
data onto existing A/V playlist documents. This can be done with a rake
task:

```bash
rake mdl_ingester:backfill_playlist_data
```

This requires an environment variable, `KALTURA_SESSION`, which provides
access to our Kaltura data. This can be generated using the following
rake task (assuming you have valid credentials for Kaltura - see Paul
Swanson for this):

```bash
rake kaltura:generate_session
```

You'll be prompted for a username and password, and the output of the
rake task must be assigned to the `KALTURA_SESSION` environment variable
on any server that's running ingestion.