Include webpack bundles in blacklight layout
To avoid conflicts with Spotlight assets, it was necessary to
conditionally include the Spotlight assets depending on whether or not
we had `content_for?(:webpack_bundles)`. However, when we are rendering
views in a regular search context, we use the Blacklight layout, so we
must also render any webpack bundles within this layout as well.