Fix schedule.rb syntax
In the DSL, `runner` is for running a code statement in the application
context. Looks like it was being used as a comment or a description
of what was to be done by the rake task we are also calling. This was
causing errors to bubble up to Sentry on the same schedule defined
in the crontab