Augment IIIF manifest to support "two-up" view in UV
UniversalViewer looks for a `viewingHint` property on the IIIF manifest
sequence in order to know whether to offer this functionality. Since it
doesn't seem to exist in the IIIF manifest we get from ContentDM, we
need to apply it ourselves on ingestion. Also provided is a Rake task to
backfill existing manifests.