Update Kaltura URLs for A/V assets
We need to use an HTTPS URL for the assets in order to reliably load
content across browsers (mixed content is handled differently in Chrome
and Firefox, for instance). Firefox will warn, but load the insecure
content anyway, whereas Chrome will upgrade the request to HTTPS
automatically. The problem there, however, is that there is no valid SSL
certificate for the Kaltura domain we had been using. By using this new
URL, we can avoid the situation entirely.