Fetch ordered playlist items from Kaltura
AV playlists with 10 or more items are out of order. They are ordered
lexographically (part 1, part 10, part 11, part 2) because in ContentDM,
the semicolon-separated items from the "audio" or "video" attributes are
ordered this way. This is likely a copy-from-Excel class of problem, but
can be solved with tighter integration with Kaltura. The Kaltura API
exposes the playlist itself, and most (if not all) of the time, we have
the ID of this entry on the ContentDM document.

In general, it seems ContentDM's "audioa" field represents the playlist
entry for both audio and video playlists, although occasionally "videoa"
is used. This change should handle either case.