Fix map marker images
This removes the version of LeafletJS sourced from the UNPKG CDN. We
have already had it as a transient gem dependency (of
blacklight-spotlight), and one outcome of having both versions was the
mangling of the map marker URLs. I've added leaflet-rails as a direct
dependency to clarify where we're sourcing it from.