Only write a crontab on the job servers
This is important because the main indexing cron job generates a JSON
blob of config data to pass to the indexing jobs. This contains the URL
for Solr, which is different on the app server than it is on the job
server (loopback vs DNS). So if we generate the config on the app server
but run the jobs on the job server, we won't have a usable URL for Solr
and the jobs will fail.