Revive left nav in Spotlight exhibit dashboards
There were some structural changes to Spotlight's layout template in
v3.0 that we missed because we have an override for that particular
template file in our project. The new template name is 'spotlight/base',
so by renaming it here, we can let the rest of Spotlight's view
machinery do its thing and the navigation will render properly.