pagination, exhibit, and zip download fixes (#275)
* Fetch images that are 800px wide when generating ZIP archive

This updates `MDL::ArchiveGenerator` to fetch the 800px wide images
rather then the full res ones.

Fixes #267

* Revive left nav in Spotlight exhibit dashboards

There were some structural changes to Spotlight's layout template in
v3.0 that we missed because we have an override for that particular
template file in our project. The new template name is 'spotlight/base',
so by renaming it here, we can let the rest of Spotlight's view
machinery do its thing and the navigation will render properly.

* Preserve offset when changing per_page pagination value

We want to keep the user at the same offset in the collection when they
change the per_page pagination settings. This sets up a redirect for
when we detect that situation.

---------

Co-authored-by: Dan Olson <olson_dan@yahoo.com>