Fix error with searching within exhibits
The search bar component template was using a route helper that wasn't
compatible with the exhibit context, and so searches were being routed
to a path that matched a route where the `/catalog` segment was
interpreted as an exhibit ID. This caused an exception because the app
was trying to look up an exhibit having the slug "catalog".

This updates the route in the template to be compatible with both the
exhibit and catalog contexts.