Use ContentDM to serve thumbnails in Download tab
Requests for thumbnails have been passing through the Rails app so that
we can queue the caching of an object's thumbnails. I'm not convinced
this cache is worthwhile, and for objects with hundreds of images, the
extra round-trip is quite expensive.

With this change, we're going to serve thumbnails directly from
ContentDM in the Download tab of citation details.