Cleanup download tab (#308)
* Combine A/V downloads with images & PDFs

This change exposes all downloadable content under the Download tab in
the citation details section. Before this, A/V content was only
downloadable via a widget in the viewer. This provides more consistency.

* Use ContentDM to serve thumbnails in Download tab

Requests for thumbnails have been passing through the Rails app so that
we can queue the caching of an object's thumbnails. I'm not convinced
this cache is worthwhile, and for objects with hundreds of images, the
extra round-trip is quite expensive.

With this change, we're going to serve thumbnails directly from
ContentDM in the Download tab of citation details.

* Lazy load thumbnails in the download tab

* Build citation download data from IIIF manifest when possible

* Simplify cite download - one download per asset