Allow for dynamic configuration of UniversalViewer
This adds a controller for rendering the UV config that can, in
principle, vary any portion of the configuration based on attirbutes of
the document being rendered. For now, it's limited to collapsing the
left panel by default for AV playlists, but it's possible to extend this
approach for other document categories as well.