Stop caching thumbnail images
The current caching strategy for thumbnail images is no longer
effective. Since we run Sidekiq jobs on a separate server than we run
our app servers, the files written to disk are not accessible to the web
processes. Furthermore, the ContentDM API is reliable enough to serve
the images without issue, so we can link directly to it for these
assets.