Update Contributing Organization handling on item details page
This change removes information from CiteDetails that's redundant with
the contributing organization's landing page. It provides a link to that
landing page, as well as a new button-styled link to a faceted search to
the organization's contributed items.

The approach for providing links to the landing pages is to scrape the
links from the site and write the results to a YAML file we store in
the repo. It's a dictionary of URLs by org name. Then as we generate the
data for the item details page, we use the dictionary to resove the URL
for the document's contributing organization. There are a few caveats to
this approach:

1) it requires the organzation names in the links on the Drupal site to
   match the values in the Solr document
2) new contributing organizations will not have links until we re-run the
   scraper
3) the scraper may fall out of date with the document structure or
   selectors of the Drupal site

A possible alternative is to store the contributing org link in the Solr
document itself, but that solution still requires maintaining a mapping
of org name to Drupal site URL somewhere, and would result in storing a
lot of duplicate data since. All in all, I feel the tradeoffs favor the
solution presented within, since it's easiest to update and maintain.