Add `identifier` field to OAI record response
This provides a new field, `identifier`, on the OAI GetRecord response.
Since the links aren't actually persisted in Solr, we need to generate
it in the `SolrDocument` constructor and merge it into the source
document that class wraps. It seems better to do this than to write a
new field to the Solr index for every document.