Fix `ActionView::Template::Error`
Sometimes, we receive requests where the query string is formatted
unexpectedly. The requester wants to filter results by some criteria,
but the way Rails parses the params produces Hash instances instead of
Strings. This has been causing rendering errors which result in a 500.
This change detects that situation and handles it gracefully.