Commit Solr index prior to notifying Sentry ETL is complete
In QA, we've occasionally seen Sentry notifications that ETL is
complete, but we won't see the expected changes reflected in Solr search
results. A Solr commit fixes it, but it should be happening already
baseed on the code in EtlAuditing. This change just swaps the order of
notification and commit, so that we only notify once the commit has
happened.