Use more specific selector in feature spec
We were seeing this test fail in CI somewhat consistently, but I
struggled to reproduce it locally. I finally discovered that I had been
working with an out of date UV config.json file, which hard coded the
left panel to be closed. As a result, I wasn't seeing the ambiguous
selector error that was present in CI. After clearing that stale file
and regenerating it with `./prep_uv.sh`, I was able to see the error and
fix the problem. The new upload artifact feature in CI helped uncover
the issue because both the screenshot and the HTML showed the left panel
open with multiple `.searchResults` selectors.