Allow users to create lists from items in the collection.
Lists are stored in IndexedDB, so they're local only to a particular
browser session. Two new routes have been added:

* `/lists` - to show all lists you've created
* `/lists/:id` - to show the contents of a specific list

A lists widget has been added to the search results page as well, to
allow for adding items found via search to a list of your choosing.
Lists can be created from this widget as well.