Alleviate resource contention in archive generation
When we download images for archive generation, we need to release HTTP
connections back to the pool (Async ecosystem), which we can do by
calling `#finish` on the response. Addionally, this introduces a
dedicated Sidekiq queue for archive generation, so we leave the default
queue open for the typical indexing responsibilities.