Fix 500 error on "missing" facet searches
When Blacklight renders its facet panels on a search results page, it
shows counts of records grouped by facet value. When it encounters a
group that has no value for that facet field, it shows a count of
"[missing]" records. Like all facet values, this is a link that further
filters the result set. However, this "missing" concept has special
representation and handling in the Blacklight library, and we have
some code in the app that was dismantling the data structure Blacklight
uses to represent a facet search value for "missing". This fixes the
problem so that these facets now work correctly.