Revive pagination on item details page
A previous attempt to prevent queries to the searches table introduced
a regression. The pagination component on the item details page happens
to depend on the presence of a `Search` object pulled from the user
session. Since we had disabled that session completely, it was no longer
rendering because the "search session" was nil. Rather than completely
nullifying it, we can instead hydrate it with a new, unpersisted
`Search` instance, which will both revive the pagination and prevent
queries to the searches table.