Add preventative measures to keep searches table from growing
We have a cron job that runs once daily to truncate the searches table.
This has apparently stopped working at some point, because our searches
table had recently grown to several Gb in size. This updates the crontab
to use the Blacklight-provided rake task instead, with the hope that it
works better than ours. It's unclear why our task would not be
successful, since it's just a simple `truncate searches` command, but
we'll try this and see how it goes.

Additionally, this adds some very naive bot checking based on user agent
that Blacklight uses to determine whether or not to save a search.