Fix UV iframe embed scenario
When we migrated UV to the CDN, we stopped serving UV assets from
/public, which ended up breaking the iframe use case. This revives
it by routing to a controller and rendering a view template. We're
using a route & view rather than a static file in /public because
we want to initialize UV with the same config options in both places
which we can do with a partial. Additionally, if we were to serve
the config as a url via a query parameter on the iframe src, UV ends
up loading the attribution pop up before the config loads, which is
not the exerience we want.