Gracefully handle 404 in IIIFManifestFormatter
We've been failing to halt processing when we get a 404 calling CDM
for a IIIF manifest. That has resulted in runtime errors because the
response body for an unsuccessful request doesn't match our expectations

This fixes the issue by halting processing when we don't get a 200 on the
manifest response.