Add GalleryDocumentComponent to serve gallery index view
Something changed in how blacklight-gallery renders its views between
v4.0.1 and v4.6.1. They seem to depend on a ViewComponents
implementation, which is inline with the overall direction of the
Blacklight libraries. We have been using the legacy view partial
overrides to customize the gallery appearance, and that is no longer a
viable solution. This introduces a GalleryDocumentComponent class that
will accommodate the expected rendering approach.

The document structure has also changed as a result of the upgrade, so
some of the CSS is no longer working. This doesn't address that issue.